make three new test
